pyproject.toml
[project]
name = "a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.4",
    "faiss-cpu>=1.13.0",
    "fastapi[all]>=0.121.2",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-google-genai>=3.0.3",
    "langchain-groq>=1.0.1",
    "langchain-openai>=1.0.3",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.3",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "pymupdf>=1.26.6",
    "pypdf>=6.3.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "typesense>=1.3.0",
]
